+++ boards/esp32c3_ble/.cargo/config.toml
@@ -15,3 +15,4 @@ MCU="esp32c3"
 ESP_IDF_VERSION = "v5.1.3"
 DEFMT_LOG="debug"
 ESP_IDF_PATH_ISSUES="warn"
+CRATE_CC_NO_DEFAULTS="1"
