+++ boards/esp32c3_ble/Cargo.toml
@@ -10,35 +10,38 @@ edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-rmk = { version = "0.1.9", path = "../../rmk", features = [
+rmk = { version = "0.1.11", path = "../../rmk", features = [
     "col2row",
     "eeprom",
 ] }
-
 esp32-nimble = { version = "0.6.0", default-features = false, features = [
     "no_std",
 ] }
-
-esp-backtrace = { version = "0.11.0", features = [
-    "esp32c3",
-    "panic-handler",
-    "exception-handler",
-    "defmt",
+esp-idf-sys = { version = "0.34", default-features = false, features = [
+    "panic_handler",
+    "alloc_handler",
 ] }
-
-esp-println = { version = "0.9.0", features = [
+esp-println = { version = "0.9", features = [
     "esp32c3",
     "defmt-espflash",
     "uart",
 ] }
-
 defmt = "0.3"
 embassy-executor = { version = "0.5.0", features = [
+    "arch-riscv32",
     "nightly",
+    "executor-thread",
     "defmt",
-    "integrated-timers",
 ] }
-embedded-hal-async = { version = "1.0.0", features = ["defmt-03"] }
+esp-idf-svc = { version = "0.48", default-features = false, features = [
+    "embassy-sync",
+    "critical-section",
+    "embassy-time-driver",
+] }
+esp-idf-hal = { verison = "*", default-features = false, features = [
+    "embassy-sync",
+] }
+riscv-rt = "0.12.2"
 
 [build-dependencies]
 xz2 = "0.1.7"
